+++ lib/departure.rb
@@ -15,14 +15,14 @@ require 'departure/errors'
 require 'departure/command'
 require 'departure/connection_base'
 require 'departure/migration'
-require 'departure/rails_integrator'
+require 'departure/rails_adapter'
 
 require 'departure/railtie' if defined?(Rails)
 
 # We need the OS not to buffer the IO to see pt-osc's output while migrating
 $stdout.sync = true
 
-Departure::RailsIntegrator.for_current.register_integrations
+Departure::RailsAdapter.for_current.register_integrations
 
 module Departure
   class << self